+++ src/CosmosDB/CosmosDB/Models/DatabaseAccount/PSDatabaseAccountKeys.cs
@@ -15,6 +15,7 @@
 namespace Microsoft.Azure.Commands.CosmosDB.Models
 {
     using System.Collections;
+    using System.Collections.Generic;
     using Microsoft.Azure.Management.CosmosDB.Models;
 
     public class PSDatabaseAccountListKeys 
@@ -43,7 +44,7 @@ namespace Microsoft.Azure.Commands.CosmosDB.Models
 
             foreach (DatabaseAccountConnectionString connectionString in databaseAccountConnectionString.ConnectionStrings)
             {
-                Keys.Add(connectionString.Description, connectionString.ConnectionString);
+                ConnectionStrings.Add(connectionString);
             }
         }
 
@@ -61,5 +62,7 @@ namespace Microsoft.Azure.Commands.CosmosDB.Models
         }
 
         public Hashtable Keys { get; set; } = new Hashtable();
+
+        public IList<DatabaseAccountConnectionString> ConnectionStrings { get; set; } = new List<DatabaseAccountConnectionString>();
     }
 }